No puedo hacer andar la impresión de páginas

+++ tp2b/4/src/eliminar_post.php
@@ -12,7 +12,9 @@ if (isset($_POST["eliminar"])) {
 		$post = new Post();
 		$post->setIdpost($_POST["idpost"]);
 		$post->setAll($post->select()[0]);
-		unlink($post->getImagen());
+		if ($post->getImagen()) {
+			unlink($post->getImagen());
+		}
 		$post->delete();
 		$resultado_eliminar = "El post se eliminó correctamente";
 	}
